requirements.txt
altgraph==0.17.5
packaging==25.0
pefile==2024.8.26
pillow==12.0.0
pyinstaller==6.17.0
pyinstaller-hooks-contrib==2025.10
PySide6==6.10.1
PySide6_Addons==6.10.1
PySide6_Essentials==6.10.1
pywin32-ctypes==0.2.3
shiboken6==6.10.1
